TST: update test org #597
